implement service

+++ backend/src/main/java/com/swe573/socialhub/dto/StatContainerDto.java
@@ -1,31 +1,31 @@
 package com.swe573.socialhub.dto;
 
 public class StatContainerDto {
-    private final int totalCount;
-    private final int last24HoursCount;
-    private final int lastWeekCount;
-    private final int lastMonthCount;
+    private final long totalCount;
+    private final long last24HoursCount;
+    private final long lastWeekCount;
+    private final long lastMonthCount;
 
-    public StatContainerDto(int totalCount, int last24HoursCount, int lastWeekCount, int lastMonthCount) {
+    public StatContainerDto(long totalCount, long last24HoursCount, long lastWeekCount, long lastMonthCount) {
         this.totalCount = totalCount;
         this.last24HoursCount = last24HoursCount;
         this.lastWeekCount = lastWeekCount;
         this.lastMonthCount = lastMonthCount;
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public int getLast24HoursCount() {
+    public long getLast24HoursCount() {
         return last24HoursCount;
     }
 
-    public int getLastWeekCount() {
+    public long getLastWeekCount() {
         return lastWeekCount;
     }
 
-    public int getLastMonthCount() {
+    public long getLastMonthCount() {
         return lastMonthCount;
     }
 }
